Move some end of day tasks to FinishUpQuest
* Open MayDay™ supply package
* Autosell Melting Junk
* Buy one-day ticket to Dinseylandfill
