+++ packages/garbo/src/index.ts
@@ -1,7 +1,6 @@
 import { Args } from "grimoire-kolmafia";
 import {
   abort,
-  availableAmount,
   buy,
   canEquip,
   cliExecute,
@@ -35,7 +34,6 @@ import {
 import {
   $class,
   $classes,
-  $coinmaster,
   $familiars,
   $item,
   $items,
@@ -89,6 +87,7 @@ import {
   CockroachSetup,
   DailyFamiliarsQuest,
   EmbezzlerFightsQuest,
+  FinishUpQuest,
   PostQuest,
   runGarboQuests,
   runSafeGarboQuests,
@@ -613,20 +612,7 @@ export function main(argString = ""): void {
           if (!targetingMeat()) runGarboQuests([EmbezzlerFightsQuest]);
           try {
             runGarboQuests([PostQuest(), ...BarfTurnQuests]);
-
-            // buy one-day tickets with FunFunds if user desires
-            if (
-              globalOptions.prefs.buyPass &&
-              availableAmount($item`FunFunds™`) >= 20 &&
-              !have($item`one-day ticket to Dinseylandfill`)
-            ) {
-              print("Buying a one-day ticket", HIGHLIGHT);
-              buy(
-                $coinmaster`The Dinsey Company Store`,
-                1,
-                $item`one-day ticket to Dinseylandfill`,
-              );
-            }
+            runGarboQuests([FinishUpQuest]);
           } finally {
             setAutoAttack(0);
           }
